# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Time Capsule services",
        "title": "API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/customevent/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new custom event with provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomEvents"
                ],
                "summary": "Add a new custom event",
                "parameters": [
                    {
                        "description": "Add custom event request",
                        "name": "AddCustomEvent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.AddCustomEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Custom event successfully added",
                        "schema": {
                            "$ref": "#/definitions/genproto.AddCustomEventResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customevent/get/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a custom event based on the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomEvents"
                ],
                "summary": "Get a custom event by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Custom event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Custom event found",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetByIdEvetsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Custom event not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customevent/gets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all custom events based on provided filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomEvents"
                ],
                "summary": "Get all custom events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID filter",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date filter",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Title filter",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Custom events found",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetAllEventsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customevent/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing custom event based on provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomEvents"
                ],
                "summary": "Update a custom event",
                "parameters": [
                    {
                        "description": "Update custom event request",
                        "name": "UpdateCustomEvent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateCustomEventsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Custom event successfully updated",
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateCustomEventsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Custom event not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customevent/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a custom event based on the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomEvents"
                ],
                "summary": "Delete a custom event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Custom event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Custom event successfully deleted",
                        "schema": {
                            "$ref": "#/definitions/genproto.DeleteCustomEventsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Custom event not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "genproto.AddCustomEventRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "genproto.AddCustomEventResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "genproto.CustomEvents": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "genproto.DeleteCustomEventsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "genproto.GetAllEventsResponse": {
            "type": "object",
            "properties": {
                "event": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.CustomEvents"
                    }
                }
            }
        },
...
        "genproto.GetByIdEvetsResponse": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/genproto.CustomEvents"
                }
            }
        },
...
        "genproto.UpdateCustomEventsRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "genproto.UpdateCustomEventsResponse": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}